Refactor outcomes
The PolicyOutcome is now an interface IPolicyOutcome with different implementing types to represent the outcome. This is so that the outcome can carry additional information like the failure message for failures and the policy name for invalid policy name outcomes.

Additionally, the result of the PolicyEvaluator.EvaluateExpression method is changed from bool to IExpressionEvaluationOutcome, which has only two implementations - PassOutcome and FailOutcome.

Also reorganized the tests and added additional tests to check for null, empty and blank expressions.